Move api to api directory
